Do not pad secret if its length is a multiple of 8
When the length of the secret is a multiple of 8, the following error
occurs:

    binascii.Error: Incorrect padding

The error occurs due to the fact that when the length of the secret is
already a multiple of 8, the padding code still adds another 8 equals
signs as padding. This fix ensures that no padding is added in this
case.